Add scripting lib

+++ render/include/rawrbox/render/stencil.hpp
@@ -161,7 +161,7 @@ namespace rawrbox {
 		// ----------
 
 		// Clip handling ----
-		std::vector<uint32_t> _clips = {};
+		std::vector<uint16_t> _clips = {};
 		// ----------
 
 		// Outline handling ----
